Revert "rm/*: ensure that -E "extra_params extra_params ..." works properly"
This reverts commit 46556b4651838433ebf6ab2388a23a3c55a16cbb.

To proper handle the extra parameters, it is decided (FD/MH) to now have
a workflow that pushes the extra parameters into a dedicated array and
manages it as an array in the final resources manager plugins.

A new mprun option will be introduced (-Z) to request for singleton
extra parameters containing space characters. Thus, the -E option of
mprun will stay compatible with current behavior by ensuring that
-E "a b c" will be translated to -E a -E b -E c automatically, resulting
in 3 different elements in the array pushed toward the RM plugins.